#156 Рефакторинг тестов с бенчами
