+++ Zilon.Core/Zilon.Core.Benchmark/BenchTests.cs
@@ -1,0 +1,24 @@
+﻿using System;
+
+using BenchmarkDotNet.Running;
+
+using NUnit.Framework;
+
+namespace Zilon.Core.Benchmark
+{
+    [TestFixture]
+    [Category("benchmark")]
+    public class BenchTests
+    {
+        [Test]
+        public void Move()
+        {
+            var buildNumber = TestContext.Parameters["buildNumber"];
+
+            Console.WriteLine($"buildNumber: {buildNumber}");
+
+            var config = new Config(buildNumber);
+            BenchmarkRunner.Run<MoveBench>(config);
+        }
+    }
+}
