+++ Zilon.Core/Zilon.Core.Benchmark/Zilon.Core.Benchmark.csproj
@@ -156,9 +156,10 @@
     <Otherwise />
   </Choose>
   <ItemGroup>
+    <Compile Include="Config.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="TheEasiestBenchmark.cs" />
-    <Compile Include="UnitTest1.cs" />
+    <Compile Include="MoveBench.cs" />
+    <Compile Include="BenchTests.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="app.config" />
